multi-agent/requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.8.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.0
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
autogen-agentchat==0.7.4
autogen-core==0.7.4
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
bond==4.1.2
bs4==0.0.2
build==1.3.0
builtin-interfaces==1.2.1
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.2.1
coloredlogs==15.0.1
composition-interfaces==1.2.1
contourpy==1.3.3
cryptography==45.0.7
curl_cffi==0.13.0
cv-bridge==3.2.1
cycler==0.12.1
dashscope==1.24.4
dataclasses-json==0.6.7
demo-nodes-py==0.20.5
diagnostic-msgs==4.8.0
diagnostic-updater==4.0.3
distro==1.9.0
domain-coordinator==0.10.0
durationpy==0.10
dwb-msgs==1.1.18
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
geometry-msgs==4.8.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
image-geometry==3.2.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
interactive-markers==2.3.2
Jinja2==3.1.6
jiter==0.10.0
joint-state-publisher==2.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.27
laser-geometry==2.4.0
launch==1.0.8
launch-ros==0.19.9
launch-testing==1.0.8
launch-testing-ros==0.19.9
launch-xml==1.0.8
launch-yaml==1.0.8
lifecycle-msgs==1.2.1
logging-demo==0.20.5
lxml==6.0.1
map-msgs==2.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mdurl==0.1.2
message-filters==4.3.7
mmh3==5.2.0
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.6.4
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.4.0
nav-2d-msgs==1.1.18
nav-msgs==4.8.0
nav2-common==1.1.18
nav2-msgs==1.1.18
nav2-simple-commander==1.0.0
numpy==2.3.3
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.107.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
ormsgpack==1.10.0
osrf-pycommon==2.1.6
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pcl-msgs==1.0.0
peewee==3.18.2
pendulum-msgs==0.20.5
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
posthog==5.4.0
propcache==0.3.2
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-qt-binding==1.1.2
pytz==2025.2
PyYAML==6.0.2
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resource-retriever==3.1.3
rich==14.1.0
rmw-dds-common==1.6.0
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpds-py==0.27.1
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.4
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9.1
sensor-msgs==4.8.0
sensor-msgs-py==4.8.0
shape-msgs==4.8.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
slam-toolbox==2.6.9
smclib==4.1.2
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sros2==0.10.6
statistics-msgs==1.2.1
std-msgs==4.8.0
std-srvs==4.8.0
stereo-msgs==4.8.0
streamlit==1.49.1
sympy==1.14.0
teleop-twist-keyboard==2.4.0
tenacity==9.1.2
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
tf2-tools==0.25.12
tiktoken==0.11.0
tokenizers==0.22.0
toml==0.10.2
topic-monitor==0.20.5
tornado==6.5.2
tqdm==4.67.1
trajectory-msgs==4.8.0
turtlesim==1.4.2
tushare==1.4.24
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
unique-identifier-msgs==2.2.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
visualization-msgs==4.8.0
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.65
zipp==3.23.0
zstandard==0.24.0
